… circuit_test.go
… registry_test.go